Se agregaron las clases que faltaban alv
