+++ Empleado.cs
@@ -15,20 +15,5 @@ namespace TrabajoGrupoGit
       return SueldoBase;
     }
   }
-    public class EmpleadoTemporal : Empleado
-    {
-        public int HorasExtra { get; set; }
-        public override double CalcularSueldo()
-        {
-            return SueldoBase + (HorasExtra * 100);
-        }
-    }
-    public class EmpleadoPermanente : Empleado
-    {
-        public double Bono { get; set; }
-        public override double CalcularSueldo()
-        {
-            return SueldoBase + Bono;
-        }
-    }
+    
 }
