+++ TrabajoGrupoGit.csproj
@@ -48,6 +48,8 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Empleado.cs" />
+    <Compile Include="EmpleadoPermanente.cs" />
+    <Compile Include="EmpleadoTemporal.cs" />
     <Compile Include="Form1.cs">
       <SubType>Form</SubType>
     </Compile>
